Refactor #75 - Wrapping TagUsageCount
