+++ app/src/main/java/com/nlab/reminder/domain/feature/home/HomeInterceptor.kt
@@ -34,7 +34,7 @@ internal class HomeInterceptor @Inject constructor(
         action<HomeAction.OnTagLongClicked> { (action) ->
             tagRepository.getUsageCount(action.tag)
                 .onSuccess { usageCount ->
-                    dispatch(HomeAction.TagConfigMetadataLoaded(action.tag, usageCount))
+                    dispatch(HomeAction.TagConfigMetadataLoaded(action.tag, usageCount.value))
                 }
                 .onFailure { e -> dispatch(HomeAction.ErrorOccurred(e)) }
                 .getOrThrow()
