+++ app/src/test/java/com/nlab/reminder/domain/feature/home/HomeStateMachineKtTest.kt
@@ -40,6 +40,7 @@ import org.mockito.kotlin.*
  */
 @OptIn(ExperimentalCoroutinesApi::class)
 class HomeStateMachineKtTest {
+    /**
     @Test
     fun `update to loading when state was init and fetch sent`() = runTest {
         testReduceTemplate(
@@ -314,5 +315,5 @@ class HomeStateMachineKtTest {
             },
             event = HomeEvent.OnTagDeleteConfirmClicked(genTag())
         )
-    }
+    }*/
 }
